+++ src/components/messages/devices/Index.vue
@@ -521,6 +521,7 @@ export default {
       } else {
         this.$store.commit(`${this.moduleName}/clear`)
       }
+      this.$store.commit(`${this.moduleName}/setSortBy`, 'timestamp')
       this.currentLimit = this.limit
       if (this.activeId) {
         let {