+++ src/store/modules/trafficViewer/actions.js
@@ -9,7 +9,10 @@ function getFromTo (val) {
 async function initTime ({ state, commit }) {
   let timestamp = Date.now()
   try {
-    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, { data: { count: 1, reverse: true } })
+    const params = { data: { count: 1, reverse: true } }
+    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, params)
+    Vue.$logger.info(`[trafficViewer]initTime`)
+    commit('reqStart', { endpoint: 'getChannelsIdentsPackets', ids: state.active, ident: state.ident, params })
     timestamp = get(resp, 'data.result[0].timestamp', Math.floor(Date.now() / 1000))
     timestamp = Math.round(timestamp * 1000)
   } catch (e) {}
@@ -23,6 +26,7 @@ async function getDevices ({ state, commit }) {
   }
   try {
     const resp = await Vue.connector.gw.getChannelsIdents(state.active, state.deviceFilter ? `*${state.deviceFilter}*` : '*')
+    commit('reqStart', { endpoint: 'getChannelsIdents', ids: state.active, ident: state.deviceFilter ? `*${state.deviceFilter}*` : '*' })
     let idents = get(resp, 'data.result', [])
     idents = idents.reduce((idents, { ident, size }) => {
       idents[ident] = { ident, size }
@@ -43,7 +47,10 @@ async function getMessages ({ state, commit }) {
   try {
     const now = Date.now()
     const to = state.to > now ? now : state.to
-    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, { data: { from: state.from / 1000, to: to / 1000, count: state.limit } })
+    const params = { data: { from: state.from / 1000, to: to / 1000, count: state.limit } }
+    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, params)
+    Vue.$logger.info(`[trafficViewer]getMessages`)
+    commit('reqStart', { endpoint: 'getChannelsIdentsPackets', ids: state.active, ident: state.ident, params })
     const messages = get(resp, 'data.result', [])
     commit('setMessages', messages)
   } catch (e) {
@@ -61,7 +68,10 @@ async function getMessagesTail ({ state, commit }) {
   try {
     const now = Date.now()
     const to = state.to > now ? now : state.to
-    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, { data: { count: state.limit, reverse: true, from: state.from / 1000, to: to / 1000 } })
+    const params = { data: { count: state.limit, reverse: true, from: state.from / 1000, to: to / 1000 } }
+    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, params)
+    Vue.$logger.info(`[trafficViewer]getMessagesTail`)
+    commit('reqStart', { endpoint: 'getChannelsIdentsPackets', ids: state.active, ident: state.ident, params })
     const messages = get(resp, 'data.result', [])
     messages.reverse()
     commit('setMessages', messages)
@@ -79,7 +89,10 @@ async function getMessagesNext ({ state, commit }) {
   let messages = []
   try {
     const from = state.messages[state.messages.length - 1].timestamp + 0.000001
-    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, { data: { count: state.limit, from, to: state.to / 1000 } })
+    const params =  { data: { count: state.limit, from, to: state.to / 1000 } }
+    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, params)
+    Vue.$logger.info(`[trafficViewer]getMessagesNext`)
+    commit('reqStart', { endpoint: 'getChannelsIdentsPackets', ids: state.active, ident: state.ident, params })
     messages = get(resp, 'data.result', [])
     commit('setMessagesAppend', messages)
   } catch (e) {
@@ -95,7 +108,10 @@ async function getMessagesPrev ({ state, commit }) {
   let messages = []
   try {
     const to = state.messages[0].timestamp - 0.000001
-    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, { data: { count: state.limit, from: state.from / 1000, to, reverse: true } })
+    const params = { data: { count: state.limit, from: state.from / 1000, to, reverse: true } }
+    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, params)
+    Vue.$logger.info(`[trafficViewer]getMessagesPrev`)
+    commit('reqStart', { endpoint: 'getChannelsIdentsPackets', ids: state.active, ident: state.ident, params })
     messages = get(resp, 'data.result', [])
     messages.reverse()
     commit('setMessagesPrepend', messages)
@@ -118,10 +134,14 @@ async function pollingGetDevices ({ state, commit }) {
     }
   })
   state.realtimeEnabled = true
+  Vue.$logger.info(`[trafficViewer]pollingGetDevices`)
+  Vue.$logger.info(`[trafficViewer]subscribe: flespi/log/gw/channels/${state.active}/connections/+/identified`)
 }
 async function removePollingGetDevices ({ state, commit }) {
   state.realtimeEnabled = false
   await Vue.connector.socket.unsubscribe(`flespi/log/gw/channels/${state.active}/connections/+/identified`)
+  Vue.$logger.info(`[trafficViewer]removePollingGetDevices`)
+  Vue.$logger.info(`[trafficViewer]unsubscribe: flespi/log/gw/channels/${state.active}/connections/+/identified`)
 }
 let mesagesPollingId = 0
 function pollingGetMessages ({ state, commit }) {
@@ -129,7 +149,10 @@ function pollingGetMessages ({ state, commit }) {
     try {
       const from = state.messages[state.messages.length - 1].timestamp + 0.000001
       const to = Math.ceil(Date.now() / 1000)
-      const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, { data: { from, to } })
+      const params = { data: { from, to } }
+      const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, params)
+      Vue.$logger.info(`[trafficViewer]pollingGetMessages`)
+      commit('reqStart', { endpoint: 'getChannelsIdentsPackets', ids: state.active, ident: state.ident, params })
       const messages = get(resp, 'data.result', [])
       commit('setMessagesAppend', messages)
     } catch (e) {
@@ -142,11 +165,15 @@ function removePollingGetMessages ({ state, commit }) {
   clearInterval(mesagesPollingId)
   mesagesPollingId = 0
   state.messagePolling = false
+  Vue.$logger.info(`[trafficViewer]removePollingGetMessages`)
 }
 async function getDevicePreview ({ state }, device) {
   let messages = []
   try {
-    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, device.ident, { data: { count: 20, reverse: true } })
+    const params = { data: { count: 20, reverse: true } }
+    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, device.ident, params)
+    Vue.$logger.info(`[trafficViewer]getDevicePreview`)
+    commit('reqStart', { endpoint: 'getChannelsIdentsPackets', ids: state.active, ident: device.ident, params })
     messages = get(resp, 'data.result', [])
   } catch (e) {}
   return messages
@@ -154,7 +181,10 @@ async function getDevicePreview ({ state }, device) {
 async function getExportData ({ state }, { from, to }) {
   let messages = []
   try {
-    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, { data: { from, to } })
+    const params = { data: { from, to } }
+    const resp = await Vue.connector.gw.getChannelsIdentsPackets(state.active, state.ident, params)
+    Vue.$logger.info(`[trafficViewer]getExportData`)
+    commit('reqStart', { endpoint: 'getChannelsIdentsPackets', ids: state.active, ident: state.ident, params })
     messages = get(resp, 'data.result', [])
   } catch (e) {}
   return messages